+++ contracts/scripts/staking/contstants.js
@@ -1,0 +1,9 @@
+const compensationData = {
+  type: 1,
+  rate: 25,
+  duration: 31104000,
+};
+
+module.exports = {
+    compensationData
+};
